requirements.txt
flask
keras
pillow
numpy